// src/types/api/bindings/UserSession.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type UserSession = { 
/**
 * The ID of the session
 */
id: string, 
/**
 * The name of the session
 */
name: string | null, 
/**
 * The ID of the user who created the session
 */
user_id: string, 
/**
 * The time the session was created
 */
created_at: string, 
/**
 * The type of session (e.g., "login", "api")
 */
type: string, 
/**
 * The time the session expires
 */
expiry: string, };